+++ internal/entity/orderItem.go
@@ -1,10 +1,10 @@
 package entity
 
 type OrderItem struct {
-	ID        int     `json:"id"`
-	OrderID   int     `json:"order_id"`
-	ProductID int     `json:"product_id"`
-	Name      string  `json:"name"`
-	Price     float64 `json:"price"`
-	Quantity  int     `json:"quantity"`
+	ID        int    `json:"id"`
+	OrderID   int    `json:"order_id"`
+	ProductID int    `json:"product_id"`
+	Name      string `json:"name"`
+	Price     int64  `json:"price"`
+	Quantity  int    `json:"quantity"`
 }
